feat: 메인 화면 디자인 변경
