+++ app/src/main/res/drawable/rectangle_green.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:padding="10dp"
+    android:shape="rectangle" >
+    <corners
+        android:bottomLeftRadius="20dp"
+        android:bottomRightRadius="20dp"
+        android:topLeftRadius="20dp"
+        android:topRightRadius="20dp" />
+
+    <gradient
+        android:startColor="#44828B"
+        android:centerColor="#49949F"
+        android:endColor="#5CAFBB"
+        android:angle="90"/>
+
+</shape>
