+++ app/src/main/res/drawable/rectangle_green2.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:padding="10dp"
+    android:shape="rectangle" >
+    <corners
+        android:bottomLeftRadius="20dp"
+        android:bottomRightRadius="20dp"
+        android:topLeftRadius="20dp"
+        android:topRightRadius="20dp" />
+
+    <gradient
+        android:startColor="#428B71"
+        android:centerColor="#499F80"
+        android:endColor="#59B493"
+        android:angle="90"/>
+
+</shape>
