+++ app/src/main/res/layout/fragment_search_shop.xml
@@ -79,9 +79,7 @@
 
     <androidx.core.widget.NestedScrollView
         android:id="@+id/resultView"
-        android:layout_marginStart="30dp"
         android:layout_marginTop="30dp"
-        android:layout_marginEnd="30dp"
         android:layout_marginBottom="70dp"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
